nal_srupymarc/examples/requirements.txt
../dist/srupymarc-2.0.0-py3-none-any.whl
tomli
pyyaml
pymarc
termcolor
lxml
